feat: fix build error
