+++ lib/consts.ts
@@ -4,11 +4,7 @@ export const MINT_FEE_RECIPIENT = "0xcfBf34d385EA2d5Eb947063b67eA226dcDA3DC38";
 export const JSON_EXTENSION_REGISTRY =
   "0xabcdefed93200601e1dfe26d6644758801d732e8";
 import { Point } from "@/types/spiral";
-import {
-  zoraCreator1155FactoryImplABI,
-  zoraCreatorFixedPriceSaleStrategyABI,
-} from "@zoralabs/protocol-deployments";
-import { encodeEventTopics } from "viem";
+
 import { baseSepolia, base } from "viem/chains";
 
 export const IS_TESTNET =
@@ -22,16 +18,6 @@ export const REFERRAL_RECIPIENT = "0x749B7b7A6944d72266Be9500FC8C221B6A7554Ce";
 // IPFS
 export const ONE_MB = 1024 * 1024;
 export const MAX_FILE_SIZE = 5 * ONE_MB;
-
-export const SETUP_NEW_CONTRACT_EVENT_SIGNATURE = encodeEventTopics({
-  abi: zoraCreator1155FactoryImplABI,
-  eventName: "SetupNewContract",
-})[0];
-
-export const MINT_COMMENT_EVENT_SIGNATURE = encodeEventTopics({
-  abi: zoraCreatorFixedPriceSaleStrategyABI,
-  eventName: "MintComment",
-})[0];
 
 export const CROSSMINT_SIGNER_ADDRESS =
   "0xa105C311fA72b8Fb78c992EcbDb8b02Ea5bd394d";
