+++ lib/viem/getViemNetwork.tsx
@@ -1,9 +1,6 @@
 import {
-  goerli,
   mainnet,
   optimism,
-  polygon,
-  polygonMumbai,
   sepolia,
   baseSepolia,
   optimismSepolia,
@@ -15,16 +12,10 @@ const getViemNetwork = (chainId: number) => {
   switch (chainId) {
     case mainnet.id:
       return mainnet;
-    case polygon.id:
-      return polygon;
     case base.id:
       return base;
     case optimism.id:
       return optimism;
-    case goerli.id:
-      return goerli;
-    case polygonMumbai.id:
-      return polygonMumbai;
     case sepolia.id:
       return sepolia;
     case baseSepolia.id:
